Update template for new `Tools` capabilities
"config.json" is now also read by the CMake project instead of just the
build helper Python script to validate and set some build parameters and
the build can be marked as "for distribution" via a Cmake cache
variable. This is set by the helper Python script if requested.

"README.md" was also updated to reflect current requirements and
features.